Edit Imports and File structure
