+++ backend/src/items/commands/create-item.command.ts
@@ -1,4 +1,4 @@
-import { CreateItemDto } from '../../interfaces/create-item.dto';
+import { CreateItemDto } from '../interfaces/create-item.dto';
 
 export class CreateItemCommand {
   constructor(public readonly createItemDto: CreateItemDto) {}
