+++ backend/src/items/commands/test-item-saga.command.ts
@@ -1,4 +1,4 @@
-import { Item } from '../../models/item.model';
+import { Item } from '../models/item.model';
 
 export class TestItemSagaCommand {
   constructor(public readonly item: Item) {}
