+++ backend/src/items/commands/test-item-saga.handler.ts
@@ -1,12 +1,13 @@
-import { UnprocessableEntityException } from '@nestjs/common';
+import { Logger, UnprocessableEntityException } from '@nestjs/common';
 import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
-import { ItemTestEvent } from '../../events/impl/item-test.event';
-import { TestItemSagaCommand } from '../impl/test-item-saga.command';
+import { ItemTestEvent } from '../events/item-test.event';
+import { TestItemSagaCommand } from './test-item-saga.command';
 
 @CommandHandler(TestItemSagaCommand)
 export class TestItemSagaHandler
   implements ICommandHandler<TestItemSagaCommand>
 {
+  private readonly logger = new Logger(TestItemSagaHandler.name);
   constructor(private readonly publisher: EventPublisher) {}
 
   // No returns, just Exceptions, rest is handled by eventHandler in CQRS
@@ -16,7 +17,7 @@ export class TestItemSagaHandler
       eventItem.apply(new ItemTestEvent(eventItem));
       eventItem.commit();
     } catch (error) {
-      console.error(error);
+      this.logger.error(error);
       throw new UnprocessableEntityException('Item could not be created');
     }
   }
