+++ backend/src/items/events/item-created.event.ts
@@ -1,4 +1,4 @@
-import { Item } from '../../models/item.model';
+import { Item } from '../models/item.model';
 
 export class ItemCreatedEvent {
   constructor(public readonly item: Item) {}
