+++ backend/src/items/events/item-test.event.ts
@@ -1,4 +1,4 @@
-import { Item } from '../../models/item.model';
+import { Item } from '../models/item.model';
 
 export class ItemTestEvent {
   constructor(public readonly item: Item) {}
