+++ backend/src/items/events/item-test.handler.ts
@@ -1,10 +1,12 @@
+import { Logger } from '@nestjs/common';
 import { EventsHandler, IEventHandler } from '@nestjs/cqrs';
-import { ItemTestEvent } from '../impl/item-test.event';
+import { ItemTestEvent } from './item-test.event';
 
 @EventsHandler(ItemTestEvent)
 export class ItemTestHandler implements IEventHandler<ItemTestEvent> {
+  private readonly logger = new Logger(ItemTestHandler.name);
   handle(event: ItemTestEvent) {
     // TODO: Remove test event and commands once explained to team
-    console.log('ItemTestHandler', event);
+    this.logger.log(event);
   }
 }
