+++ backend/src/items/queries/get-item.handler.ts
@@ -1,12 +1,14 @@
-import { NotFoundException } from '@nestjs/common';
+import { Logger, NotFoundException } from '@nestjs/common';
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Item } from '../../models/item.model';
-import { GetItemQuery } from '../impl';
+import { Item } from '../models/item.model';
+import { GetItemQuery } from './get-item.query';
 
 @QueryHandler(GetItemQuery)
 export class GetItemHandler implements IQueryHandler<GetItemQuery> {
+  private readonly logger = new Logger(GetItemHandler.name);
+
   constructor(
     @InjectRepository(Item)
     private repository: Repository<Item>,
@@ -14,9 +16,9 @@ export class GetItemHandler implements IQueryHandler<GetItemQuery> {
 
   async execute(query: GetItemQuery) {
     const result = await this.repository.findOneBy(query);
-    if (!result)
+    if (!result) {
       throw new NotFoundException(`Item with slug ${query.slug} not found`);
-
+    }
     // TODO: Can you return new ...Exception... and it throws it later? => return result || new NotFoundException(`Item with id ${query.id} not found`);
     return result;
   }
