+++ backend/src/items/queries/get-item.query.ts
