--- backend/src/items/queries/impl/index.ts
@@ -1,1 +1,0 @@
-export * from './get-item.query';
